+++ dart/geobase/lib/geodesy.dart
@@ -1,14 +1,15 @@
 /// Ellipsoidal (*vincenty*) and spherical (*great circle*, *rhumb line*)
-/// geodesy tools.
+/// geodesy tools, with UTM and MGRS coordinate conversions.
 ///
-/// Spherical and ellipsoidal geodesy tools by Chris Veness 2002-2024 (MIT
-/// Licence) ported to Dart by Navibyte.
+/// Geodesy tools by Chris Veness 2002-2024 (MIT Licence) ported to Dart by
+/// Navibyte.
 ///
 /// See links for the original work:
 /// * https://github.com/chrisveness/geodesy
-/// * www.movable-type.co.uk/scripts/latlong.html
-/// * www.movable-type.co.uk/scripts/geodesy-library.html#latlon-spherical
+/// * https://www.movable-type.co.uk/scripts/latlong.html
+/// * https://www.movable-type.co.uk/scripts/geodesy-library.html#latlon-spherical
 /// * https://www.movable-type.co.uk/scripts/latlong-vincenty.html
+/// * https://www.movable-type.co.uk/scripts/latlong-utm-mgrs.html  
 ///
 /// This libary exports a subset of `package:geobase/geobase.dart`.
 ///
@@ -21,5 +22,6 @@ export 'src/geodesy/ellipsoidal/ellipsoidal.dart';
 export 'src/geodesy/ellipsoidal/ellipsoidal_extension.dart';
 export 'src/geodesy/ellipsoidal/ellipsoidal_vincenty.dart';
 export 'src/geodesy/ellipsoidal/utm.dart';
+export 'src/geodesy/ellipsoidal/utm_mgrs.dart';
 export 'src/geodesy/spherical/spherical_great_circle.dart';
 export 'src/geodesy/spherical/spherical_rhumb_line.dart';